+++ knarr-server/src/main/java/edu/gmu/swe/knarr/server/concolic/picker/MaxStaticCoveragePicker.java
@@ -45,6 +45,6 @@ public class MaxStaticCoveragePicker extends Picker {
 
     @Override
     public void score(Input in) {
-        in.score = in.coverage.countCodeCoverage();
+        in.score = (int) in.coverage.countCodeCoverage();
     }
 }
